🦄 refactor(icon): 移除之前的Icon组件及依赖
