+++ src/components/Icon/_style.scss
@@ -1,5 +1,20 @@
-@each $key, $val in $theme-colors {
-  .icon-#{$key} {
-    color: $val;
-  }
+@mixin loading-animation {
+    @keyframes loading {
+        100% {
+            transform: rotate(360deg);
+        }
+    }
+
+    @-webkit-keyframes loading {
+        100% {
+            -webkit-transform: rotate(360deg);
+            transform: rotate(360deg);
+        }
+    }
+}
+
+.icon-loading {
+    display: inline-block;
+    animation: loading 1s linear infinite;
+    @include loading-animation;
 }
