--- src/components/Icon2/_style.scss
@@ -1,20 +1,0 @@
-@mixin loading-animation {
-    @keyframes loading {
-        100% {
-            transform: rotate(360deg);
-        }
-    }
-
-    @-webkit-keyframes loading {
-        100% {
-            -webkit-transform: rotate(360deg);
-            transform: rotate(360deg);
-        }
-    }
-}
-
-.icon-loading {
-    display: inline-block;
-    animation: loading 1s linear infinite;
-    @include loading-animation;
-}
