+++ src/styles/index.scss
@@ -8,5 +8,4 @@
 @import '../components/Alert/style';
 @import '../components//Menu/styles';
 @import '../components/Icon/style';
-@import '../components/Icon2/style';
 @import '../components/Message/style';
